add reducers
