+++ src/reducers/chat.js
@@ -1,0 +1,11 @@
+import { combineReducers } from 'redux'
+import messages from './messages'
+import users from './users'
+
+
+const chat = combineReducers({
+  messages,
+  users
+})
+
+export default chat
