+++ src/reducers/messages.js
@@ -1,0 +1,17 @@
+const messages = (state = [], action) => {
+  switch (action.type) {
+    case 'ADD_MESSAGE':
+    case 'MESSAGE_RECEIVED':
+    return state.concat([
+      {
+        message: action.message,
+        author: action.author,
+        id: action.id
+      }
+    ])
+    default:
+      return state
+  }
+}
+
+export default messages
